react file restructure
