--- src/components/Login.tsx
@@ -1,9 +1,0 @@
-const Login: React.FC = () => {
-    return (
-        <div>
-            
-        </div>
-    );    
-}
-
-export default Login;
