+++ src/components/StorePlaceholder.tsx
